checkpoint: moved most websocket-specific code to websockets.js
