+++ lib/unblocker.js
@@ -111,13 +111,6 @@ function Unblocker(config) {
   var getRealUrl = Unblocker.getRealUrl(config);
 
   function handleRequest(clientRequest, clientResponse, next) {
-    // default to express's more advanced version of this when available (handles X-Forwarded-Protocol headers)
-    if (!clientRequest.protocol) {
-      clientRequest.protocol = clientRequest.connection.encrypted
-        ? "https"
-        : "http";
-    }
-
     // convenience methods
     clientRequest.thisHost = thisHost.bind(thisHost, clientRequest);
     clientRequest.thisSite = thisSite.bind(thisSite, clientRequest);
@@ -225,8 +218,14 @@ function Unblocker(config) {
 
   // returns the http://site.com/proxy
   function thisSite(request) {
-    // todo: return https when appropriate
-    return request.protocol + "://" + thisHost(request) + config.prefix;
+    // default to express's more advanced version of this when available (handles X-Forwarded-Protocol headers)
+    const proto =
+      request.protocol ||
+      request.headers["X-Forwarded-Protocol"] ||
+      request.connection.encrypted
+        ? "https"
+        : "http";
+    return proto + "://" + thisHost(request) + config.prefix;
   }
 
   function redirectTo(request, response, site, headers) {
@@ -255,10 +254,8 @@ function Unblocker(config) {
   }
 
   // websocket support
-  // todo: move most or all of this to websockets.js
-
-  var http = require("http"),
-    https = require("https");
+  // todo: dedupe
+  const proxyWebsocket = Unblocker.websockets(config);
 
   handleRequest.onUpgrade = function onUpgrade(
     clientRequest,
@@ -266,11 +263,6 @@ function Unblocker(config) {
     clientHead
   ) {
     debugWS("handling websocket req to", clientRequest.url);
-    // default to express's more advanced version of this when available (handles X-Forwarded-Protocol headers)
-    clientRequest.protocol =
-      clientRequest.protocol || clientRequest.connection.encrypted
-        ? "https"
-        : "http";
 
     // convenience methods
     clientRequest.thisHost = thisHost.bind(thisHost, clientRequest);
@@ -287,100 +279,16 @@ function Unblocker(config) {
       // this is how api consumers can hook into requests. The data object is passed to all requestMiddleware before the request is sent to the remote server, and it is passed through all responseMiddleware before being sent to the client.
       var data = {
         url: formatted,
+        uri,
         clientRequest: clientRequest,
-        clientResponse: clientSocket,
         clientSocket: clientSocket,
         headers: _.cloneDeep(clientRequest.headers),
         stream: clientRequest,
+        isWebsocket: true,
+        clientHead,
       };
 
-      var onError = function () {
-        console.error("error", arguments);
-        clientSocket.end();
-      };
-
-      var middlewareHandledRequest = _.some(
-        config.requestMiddleware,
-        function (middleware) {
-          middleware(data);
-          return data.clientResponse.headersSent; // if true, then _.some will stop processing middleware here because we can no longer
-        }
-      );
-
-      if (!middlewareHandledRequest) {
-        var options = {
-          host: uri.hostname,
-          port: uri.port,
-          path: uri.path,
-          method: data.clientRequest.method,
-          headers: data.headers,
-        };
-
-        //set the agent for the request.
-        if (uri.protocol == "http:" && config.httpAgent) {
-          options.agent = config.httpAgent;
-        }
-        if (uri.protocol == "https:" && config.httpsAgent) {
-          options.agent = config.httpsAgent;
-        }
-
-        // what protocol to use for outgoing connections.
-        var proto = uri.protocol == "https:" ? https : http;
-
-        debugWS("sending remote request: ", options);
-
-        data.remoteRequest = proto.request(options, function (remoteResponse) {
-          debugWS("websocket remote response recieved");
-          data.remoteResponse = remoteResponse;
-          data.remoteResponse.on("error", onError);
-        });
-
-        data.remoteRequest.on("error", onError);
-
-        if (clientHead && clientHead.length) {
-          debugWS("sending clientHead", clientHead);
-          data.remoteRequest.write(clientHead);
-        }
-
-        data.remoteRequest.end(); // Done sending opening data. Doesn't prevent upgrad event or close the connection.
-
-        data.remoteRequest.on(
-          "upgrade",
-          function (remoteResponse, remoteSocket, remoteHead) {
-            debugWS("websocket proxy established", remoteResponse.rawHeaders);
-
-            var key = true;
-            var headers = "HTTP/1.1 101 Web Socket Protocol Handshake\r\n";
-            remoteResponse.rawHeaders.forEach(function (val) {
-              headers += val + (key ? ": " : "\r\n");
-              key = !key;
-            });
-            headers += "\r\n";
-
-            debugWS("sending headers", headers);
-
-            clientSocket.write(headers);
-
-            data.remoteResponse = remoteResponse;
-
-            if (remoteHead && remoteHead.length) {
-              debugWS("sending remoteHead", remoteHead);
-              clientSocket.write(remoteHead);
-            }
-            clientSocket.pipe(remoteSocket);
-            remoteSocket.pipe(clientSocket);
-
-            // data dump
-            clientSocket.on("data", function (chunk) {
-              debugWS("from ws client: ", chunk.toString());
-            });
-            remoteSocket.on("data", function (chunk) {
-              debugWS("from ws remote: ", chunk.toString());
-            });
-            // todo: handle errors, clean exits, etc
-          }
-        );
-      }
+      proxyWebsocket(data);
     } else {
       // any other url gets redirected to the correct proxied url if we can
       // determine it based on their referrer, or passed back to express (or whatever) otherwise
